refactor : FeedApi Service 구조 변경
